+++ Cargo.toml
@@ -5,6 +5,7 @@ edition = "2024"
 
 [dependencies]
 yojo-art-utils = { path="./utils" }
+yojo-art-models = { path="./models" }
 tokio-stream = "*"
 tokio = { version = "1.0", features = ["rt-multi-thread","signal","process"] }
 tokio-util = { version = "0.7.8", features = ["io"] }
