+++ models/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "yojo-art-models"
+version = "0.1.0"
+authors = ["kozakura <kozakura@kzkr.xyz>"]
+edition = "2024"
+publish = false
+
+[lib]
+
+[dependencies]
+yojo-art-utils = { path="../utils" }
+serde = {version="^1.0",features=["derive"]}
+chrono = {version="0.4",features=["serde"]}
+serde_json ="1"
+diesel ={version="*", features = ["chrono","numeric","serde_json","64-column-tables"]}
+diesel-async={version="0.5.2", features = ["postgres","bb8"]}
+diesel-derive-enum ={version="2.1.0", features = ["postgres"]}
+strum_macros = "0.26.4"
+strum = "0.26.3"
