+++ models/src/announcement.rs
