+++ models/src/announcement_read.rs
