+++ models/src/avatar_decoration.rs
