+++ models/src/emoji.rs
